fix: cast to silence compiler warning for Qt6 without ksyntaxhighlighting
Fixes:
```
/__w/hotspot/hotspot/src/models/highlightedtext.cpp: In member function 'virtual LineFormat HighlightingImplementation::formatLine(const QString&)':
/__w/hotspot/hotspot/src/models/highlightedtext.cpp:149:58: error: narrowing conversion of '(& line)->QString::length()' from 'qsizetype' {aka 'long long int'} to 'int' [-Werror=narrowing]
  149 |         return {{QTextLayout::FormatRange {0, line.length(), {}}}};
      |                                               ~~~~~~~~~~~^~
```

+++ src/models/highlightedtext.cpp
@@ -146,7 +146,7 @@ private:
     // stub implementation necessary for testing
     virtual LineFormat formatLine(const QString& line)
     {
-        return {{QTextLayout::FormatRange {0, line.length(), {}}}};
+        return {{QTextLayout::FormatRange {0, static_cast<int>(line.length()), {}}}};
     }
 
     Q_DISABLE_COPY(HighlightingImplementation)
